Added basic mtr-ui directory
This will be the basic structure of the mtr web app.  It's based on the geonet-web/origin
layout which uses bootstrap for it's templates.

Half way through adding a page that will display a table for all matching tags.

This is the first of the handlers that will use the mtr-api directly.  For now it's
just creating a table.  Later it will embed svg plots and other magic.